Fix test failure and move test code for auth to `authentication_test.go`
